Address comments from JINMEI, Tatuya
... an application cannot
always/fully control a specific TCP segment contains how much data: it
depends on the state of the TCP connection, specific implementation
details of the TCP stack, etc.  I suggest rephrasing this as follows

  TCP does not suffer from UDP's issues with fragmentation.

Exactly what does this mean?  Is this to say IP fragmentation won't
happen for TCP?  Or does this mean even if fragmentation occurs TCP is
free from the "UDP's issues" (what are they, btw)?